requirements.txt
python-binance>=1.0.16
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
scikit-learn>=1.0.0
joblib>=1.1.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
